# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Open ICD Coder API",
  "description": "Prototype backend for AI-augmented medical coding: ICD-10-CM terminology search (Stage 1) and the coding pipeline \u2014 candidates, decision engine, human review (Stages 2-4). See docs/goal.md.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/terminology/search": {
   "get": {
    "tags": [
     "Terminology"
    ],
    "summary": "Search codes by code or text",
    "description": "Code-shaped queries ('K35', 'k3530', 'K35.3') always return codes with that prefix in tabular order. Otherwise `mode` picks the method:\n\n- **text** \u2014 PostgreSQL full-text search (word stems, any order) + trigram similarity (typos) over the description and inclusion terms.\n- **semantic** \u2014 nearest codes by embedding similarity (pgvector); finds lay wording like 'heart attack'. Needs the embedding server.\n- **hybrid** \u2014 both rankings merged with reciprocal rank fusion; each hit reports its `text_rank` and `semantic_rank`.",
    "operationId": "search_terminology_concepts",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 1,
       "maxLength": 200,
       "description": "e.g. 'acute appendicitis with localized peritonitis', 'heart attack' or 'K35.3'",
       "title": "Q"
      },
      "description": "e.g. 'acute appendicitis with localized peritonitis', 'heart attack' or 'K35.3'"
     },
     {
      "name": "mode",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "text",
        "semantic",
        "hybrid"
       ],
       "type": "string",
       "description": "text | semantic | hybrid (default from configs/config.yaml)",
       "default": "hybrid",
       "title": "Mode"
      },
      "description": "text | semantic | hybrid (default from configs/config.yaml)"
     },
     {
      "name": "system",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Code system canonical URL",
       "default": "http://hl7.org/fhir/sid/icd-10-cm",
       "title": "System"
      },
      "description": "Code system canonical URL"
     },
     {
      "name": "version",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Release, e.g. '2027'. Default: latest loaded",
       "title": "Version"
      },
      "description": "Release, e.g. '2027'. Default: latest loaded"
     },
     {
      "name": "billable_only",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "Exclude non-billable category headers",
       "default": false,
       "title": "Billable Only"
      },
      "description": "Exclude non-billable category headers"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 20,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchResponse"
        }
       }
      }
     },
     "503": {
      "description": "Embedding server unreachable (semantic/hybrid)"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/terminology/concepts/{code}": {
   "get": {
    "tags": [
     "Terminology"
    ],
    "summary": "Get one code with its hierarchy and notes",
    "description": "Accepts dotted or undotted ICD-10-CM codes ('K35.30' or 'K3530'). Returns the parent, direct children, inclusion terms and tabular notes (excludes1/2, code first, use additional code, 7th character), plus `inherited_notes` from ancestor categories, which apply to this code too.",
    "operationId": "get_terminology_concept",
    "parameters": [
     {
      "name": "code",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Code"
      }
     },
     {
      "name": "system",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Code system canonical URL",
       "default": "http://hl7.org/fhir/sid/icd-10-cm",
       "title": "System"
      },
      "description": "Code system canonical URL"
     },
     {
      "name": "version",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Release, e.g. '2027'. Default: latest loaded",
       "title": "Version"
      },
      "description": "Release, e.g. '2027'. Default: latest loaded"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ConceptDetailResponse"
        }
       }
      }
     },
     "404": {
      "description": "Code or code system not found"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ConceptDetailResponse": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "display": {
      "type": "string",
      "title": "Display"
     },
     "short_display": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Short Display"
     },
     "is_billable": {
      "type": "boolean",
      "title": "Is Billable"
     },
     "system": {
      "type": "string",
      "title": "System"
     },
     "version": {
      "type": "string",
      "title": "Version"
     },
     "parent": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ConceptSummary"
       },
       {
        "type": "null"
       }
      ]
     },
     "children": {
      "items": {
       "$ref": "#/components/schemas/ConceptSummary"
      },
      "type": "array",
      "title": "Children"
     },
     "synonyms": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Synonyms"
     },
     "notes": {
      "additionalProperties": {
       "items": {
        "type": "string"
       },
       "type": "array"
      },
      "type": "object",
      "title": "Notes"
     },
     "inherited_notes": {
      "items": {
       "$ref": "#/components/schemas/InheritedNotes"
      },
      "type": "array",
      "title": "Inherited Notes"
     }
    },
    "type": "object",
    "required": [
     "code",
     "display",
     "is_billable",
     "system",
     "version",
     "children",
     "synonyms",
     "notes",
     "inherited_notes"
    ],
    "title": "ConceptDetailResponse"
   },
   "ConceptSummary": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "display": {
      "type": "string",
      "title": "Display"
     },
     "short_display": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Short Display"
     },
     "is_billable": {
      "type": "boolean",
      "title": "Is Billable"
     }
    },
    "type": "object",
    "required": [
     "code",
     "display",
     "is_billable"
    ],
    "title": "ConceptSummary"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "InheritedNotes": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "display": {
      "type": "string",
      "title": "Display"
     },
     "notes": {
      "additionalProperties": {
       "items": {
        "type": "string"
       },
       "type": "array"
      },
      "type": "object",
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "code",
     "display",
     "notes"
    ],
    "title": "InheritedNotes",
    "description": "Notes declared on an ancestor category, which apply to this code too."
   },
...
   "SearchHit": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "display": {
      "type": "string",
      "title": "Display"
     },
     "short_display": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Short Display"
     },
     "is_billable": {
      "type": "boolean",
      "title": "Is Billable"
     },
     "score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Score"
     },
     "text_rank": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Text Rank"
     },
     "semantic_rank": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Semantic Rank"
     },
     "matched_on": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Matched On"
     },
     "variants": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Variants",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "code",
     "display",
     "is_billable"
    ],
    "title": "SearchHit"
   },
   "SearchResponse": {
    "properties": {
     "system": {
      "type": "string",
      "title": "System"
     },
     "version": {
      "type": "string",
      "title": "Version"
     },
     "query": {
      "type": "string",
      "title": "Query"
     },
     "match_type": {
      "type": "string",
      "enum": [
       "code",
       "text",
       "semantic",
       "hybrid"
      ],
      "title": "Match Type"
     },
     "total": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total"
     },
     "limit": {
      "type": "integer",
      "title": "Limit"
     },
     "offset": {
      "type": "integer",
      "title": "Offset"
     },
     "data": {
      "items": {
       "$ref": "#/components/schemas/SearchHit"
      },
      "type": "array",
      "title": "Data"
     }
    },
    "type": "object",
    "required": [
     "system",
     "version",
     "query",
     "match_type",
     "total",
     "limit",
     "offset",
     "data"
    ],
    "title": "SearchResponse"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}